zebra: allow removal of an alias
This change solves a problem which could prevent an alias from being
removed by 'no ip address ...' command.

On FreeBSD, an alias has the same prefix with other addresses
on the same interface, not /32.  In this case, !CONNECTED_PEER(ifc) and
a non-NULL d can happen at the same time.